systemd: improve service file
use Type=simple and no fork to avoid PIDFile race

add Restart=on-abnormal

This is a long-running process that once started by systemd should
stay running forever. In the event of a fatal error and unclean exit,
we want systemd to restart the program. We use on-abnormal because
there are valid nonzero exit codes (EX_USAGE, EX_OSERR, EX_CONFIG)
which the program can properly return indicating a configuration or
command line error, or inability to allocate memory early on. In these
events, restarting would be inappropriate, but on-failure would try
anyhow.

https://github.com/trusteddomainproject/OpenARC/pull/154