Improve handling of ARC state
Validation of the state is a part of sealing and should not be skipped
even in plain "s" mode.

Always record the current state in ARC-Authentication-Results.

Allowing the most recent Authentication-Results from the ADMD to
override the chain state is fine when explicitly requested and if it's
failing due to message modification, but it can't resurrect a dead
chain.

It also shouldn't force a failure just because there are multiple
A-R headers, it should ignore the excess ones.