clang-format
I tried for a while to come up with a config that perfectly matched
the existing style and the closest I could get changed about 15% of
the lines and looked bad on moderately indented blocks because of
the use of tabs. Moving away from tabs changes more lines, but looks
better.